doc technique et resolutions de derniers details

+++ src/Controller/CustomerController.php
@@ -9,12 +9,47 @@ use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 class CustomerController extends AbstractController
 {
+
+    /**
+     * Liste les users pour un customer spécifique
+     * @OA\Response(
+     *     response=200,
+     *     description="Retourne la liste des users pour un customer",
+     *     @OA\JsonContent(
+     *        type="array",
+     *        @OA\Items(ref=@Model(type=User::class))
+     *     )
+     * )
+     * @OA\Response(
+     *      response=401,
+     *      description="JWT Token not found",
+     * )
+     * @OA\Response(
+     *      response=403,
+     *      description="Vous n'avez pas accès a ce customer.",
+     * )
+     * @OA\Parameter(
+     *     name="page",
+     *     in="query",
+     *     description="Le page actuel d'affichage des users",
+     *     @OA\Schema(type="integer")
+     * )
+     * @OA\Parameter(
+     *      name="limit",
+     *      in="query",
+     *      description="Le nombre d'affichage des users",
+     *      @OA\Schema(type="integer")
+     *  )
+     *
+     * @OA\Tag(name="users")
+     */
     #[Route('/api/customer/{id}', name: 'get_customers', methods:['get'])]
     public function index(ManagerRegistry $doctrine , int $id, UserInterface $authenticatedUser, Request $request, PaginatorInterface $paginator): JsonResponse
     {
@@ -24,7 +59,7 @@ class CustomerController extends AbstractController
 
 
         if ($authenticatedUser->getId() !== $id) {
-            $errorMessage = 'You do not have access to this customer.';
+            $errorMessage = 'Vous n\'avez pas accès a ce customer.';
 
             return new JsonResponse(['error' => $errorMessage], Response::HTTP_FORBIDDEN);
         }
@@ -37,11 +72,6 @@ class CustomerController extends AbstractController
             ->findUsersByCustomer($id);
 
 
-
-        if (!$users) {
-
-            return $this->json('No customer found for id ' . $id, 404);
-        }
 
         $pagination = $paginator->paginate(
             $users, // Sur quoi on pagine
@@ -65,12 +95,36 @@ class CustomerController extends AbstractController
 
     }
 
-    #[Route('/api/customer/{idCustomer}/user/{idUser}', name: 'get_a_customer')]
+    /**
+     * Retourne un user spécifique à partir d'un customer
+     * @OA\Response(
+     *     response=200,
+     *     description="Retourne le user",
+     *     @OA\JsonContent(
+     *        @Model(type=User::class)
+     *     )
+     * )
+     * @OA\Response(
+     *      response=401,
+     *      description="JWT Token not found",
+     * )
+     * @OA\Response(
+     *       response=403,
+     *       description="Vous n'avez pas accès a ce customer.",
+     *  )
+     * @OA\Response(
+     *       response=404,
+     *       description="Cette ressource n'existe pas.",
+     *  )
+     *
+     * @OA\Tag(name="user")
+     */
+    #[Route('/api/customer/{idCustomer}/user/{idUser}', name: 'get_a_customer', methods:['get'])]
     public function getAUserfromCustomerId(ManagerRegistry $doctrine , int $idCustomer, int $idUser, UserInterface $authenticatedUser): JsonResponse
     {
 
         if ($authenticatedUser->getId() !== $idCustomer) {
-            $errorMessage = 'You do not have access to this customer.';
+            $errorMessage = 'Vous n\'avez pas accès a ce customer.';
 
             return new JsonResponse(['error' => $errorMessage], Response::HTTP_FORBIDDEN);
         }
@@ -83,6 +137,11 @@ class CustomerController extends AbstractController
                 'customer' => $idCustomer,
             ]);
 
+        if($user == null){
+            $errorMessage = 'Cette ressource n\'existe pas.';
+            return new JsonResponse(['error' => $errorMessage], Response::HTTP_NOT_FOUND);
+        }
+
         $data =  [
             'id' => $user->getId(),
             'name' => $user->getName(),
@@ -93,11 +152,33 @@ class CustomerController extends AbstractController
 
     }
 
-    #[Route('/api/customer/{idCustomer}/deleteUser/{idUser}', name: 'delete_a_customer')]
+
+    /**
+     * Supprimer un user spécifique à partir d'un customer
+     * @OA\Response(
+     *     response=204,
+     *     description="Supprimer le user",
+     *     content: null
+     * )
+     * @OA\Response(
+     *      response=401,
+     *      description="JWT Token not found",
+     * )
+     * @OA\Response(
+     *       response=403,
+     *       description="Vous n'avez pas accès a ce customer.",
+     *  )
+     * @OA\Response(
+     *        response=404,
+     *        description="Cette ressource n'existe pas.",
+     *   )
+     * @OA\Tag(name="deleteUser")
+     */
+    #[Route('/api/customer/{idCustomer}/user/{idUser}', name: 'delete_user', methods:['DELETE'])]
     public function deleteUser(ManagerRegistry $doctrine , int $idCustomer, int $idUser, UserInterface $authenticatedUser): JsonResponse
     {
         if ($authenticatedUser->getId() !== $idCustomer) {
-            $errorMessage = 'You do not have access to this customer.';
+            $errorMessage = 'Vous n\'avez pas accès a ce customer.';
 
             return new JsonResponse(['error' => $errorMessage], Response::HTTP_FORBIDDEN);
         }
@@ -109,24 +190,46 @@ class CustomerController extends AbstractController
                 'customer' => $idCustomer,
             ]);
 
+        if($user == null){
+            $errorMessage = 'Cette ressource n\'existe pas.';
+            return new JsonResponse(['error' => $errorMessage], Response::HTTP_NOT_FOUND);
+        }
         $entityManager->remove($user);
 
         $entityManager->flush();
 
-        $users = $entityManager->getRepository(Customer::class)
-            ->findOneBy([
-                'id' => $idUser
-            ]);
 
-        return $this->json($users);
+        return $this->json(['message' => 'Suppression exécutée avec succès'], 204);
     }
 
 
-    #[Route('/api/customer/{idCustomer}/create', name: 'add_a_customer', methods:['POST'])]
+    /**
+     * Creer un user à partir d'un customer spécifique
+     * @OA\RequestBody (
+     *      description="Ajout d'un user",
+     *      @OA\JsonContent(
+     *         @Model(type=User::class)
+     *      )
+     *  )
+     * @OA\Response(
+     *     response=201,
+     *     description="Ajout du user avec succès"
+     * )
+     * @OA\Response(
+     *      response=401,
+     *      description="JWT Token not found",
+     * )
+     * @OA\Response(
+     *       response=403,
+     *       description="Vous n'avez pas accès a ce customer.",
+     *  )
+     * @OA\Tag(name="createUser")
+     */
+    #[Route('/api/customer/{idCustomer}', name: 'add_user', methods:['POST'])]
     public function addUser(ManagerRegistry $doctrine , int $idCustomer, Request $request,  UserInterface $authenticatedUser): JsonResponse
     {
         if ($authenticatedUser->getId() !== $idCustomer) {
-            $errorMessage = 'You do not have access to this customer.';
+            $errorMessage = 'Vous n\'avez pas accès a ce customer.';
 
             return new JsonResponse(['error' => $errorMessage], Response::HTTP_FORBIDDEN);
         }
@@ -138,7 +241,6 @@ class CustomerController extends AbstractController
         //dd($content);
         $user->setName($content['name']);
         $user->setEmail($content['email']);
-        $user->setPassword($content['password']);
         
 
         $customer = $entityManager->getRepository(Customer::class)
@@ -152,6 +254,6 @@ class CustomerController extends AbstractController
         $entityManager->flush();
 
         // Return a response, you can customize this based on your needs
-        return $this->json(['message' => 'User created successfully'], 201);
+        return $this->json(['message' => 'Ajout du user avec succès'], 201);
     }
 }
